New approach of the analysis of the comments
Now all the comments are concatenated at the end of the analysis of the closed feedback, the resulting paragraph is given to ChatGPT to do the classifications. Now the maximum amount of queryes performed per user are only 2: for clasifying the comments and for creating the tasks. Also the strengths and weaknesses of the user are taken from all the feedback.